Add missing keywords and fixed numeric literals

+++ core/src/main/scala/de/leanovate/jbj/ast/stmt/ContinueStmt.scala
@@ -3,6 +3,6 @@ package de.leanovate.jbj.ast.stmt
 import de.leanovate.jbj.ast.{FilePosition, Stmt}
 import de.leanovate.jbj.runtime.{ContinueExecResult, Context}
 
-case class ContinueStmt(position: FilePosition, depth: Int) extends Stmt {
+case class ContinueStmt(position: FilePosition, depth: Long) extends Stmt {
   def exec(ctx: Context) = ContinueExecResult(depth)
 }
