+++ core/src/main/scala/de/leanovate/jbj/runtime/ExecResult.scala
@@ -7,6 +7,6 @@ case class SuccessExecResult() extends ExecResult
 
 case class ReturnExecResult(value: Value) extends ExecResult
 
-case class BreakExecResult(depth: Int) extends ExecResult
+case class BreakExecResult(depth: Long) extends ExecResult
 
-case class ContinueExecResult(depth: Int) extends ExecResult
+case class ContinueExecResult(depth: Long) extends ExecResult
